+++ packages/client/src/components/Game/index.ts
@@ -1,0 +1,2 @@
+import { Game } from './Game'
+export { Game }
