+++ packages/client/src/domain/Game/AnimateStrategy.ts
