+++ packages/client/src/domain/Game/Collectable.ts
