+++ packages/client/src/domain/Game/RenderStrategy.ts
