+++ packages/client/src/domain/Game/interfaces.ts
