+++ packages/client/src/pages/Game/Game.tsx
@@ -1,7 +1,12 @@
 import React from 'react'
+import { Game as GameComponent } from '../../components/Game'
 
 const Game: React.FC = function () {
-  return <div>Game</div>
+  return (
+    <div>
+      <GameComponent />
+    </div>
+  )
 }
 
 export { Game }
